mobile-notes-agent/requirements.txt
anthropic>=0.42.0
python-telegram-bot>=20.7
apscheduler>=3.10.4
python-dotenv>=1.0.0
